Improve setting option values, and add user events
Add user events to allow the user to customize the indentation after the
detection, to fix some options that might be inherited from the initial
configuration. Allows to customize via the configuration which options
are set.

When tabs are detected, set shiftwidth=0 to ensure that the shifting
value is the same as the tabstop. Same with softtabstop=0 to disable it.
When spaces are detected, make optional setting of the tabstop and
softtabstop option.